client/requirements.txt
flux_led==0.26.11
websocket-client==1.2.1